refactor(games): Move stats calculation to own Concern
