+++ app/models/games/player_stats.rb
@@ -1,0 +1,5 @@
+module Games
+  class PlayerStats
+    attr_accessor :player, :win_percent
+  end
+end
